[TCPIP] Fix listen backlog u8 truncation
DispTdiListen requests a backlog of 1024, but TCPListen passed it into
LibTCPListen through a u8_t parameter and message field, truncating it
to 0. lwIP's tcp_backlog_set then coerced 0 to 1, so every listening
socket ran with an effective backlog of 1: with TCP_LISTEN_BACKLOG
enabled, tcp_listen_input silently drops any SYN while one handshake is
already in flight, causing connect-refused storms under concurrent
inbound connections.

Carry the backlog as ULONG through TCPListen/LibTCPListen and the
tcpip-thread message, and clamp it to lwIP's u8_t range (255) only at
the tcp_listen_with_backlog call.

+++ drivers/network/tcpip/ip/transport/tcp/accept.c
@@ -48,7 +48,7 @@ NTSTATUS TCPCheckPeerForAccept(PVOID Context,
 
 /* This listen is on a socket we keep as internal.  That socket has the same
  * lifetime as the address file */
-NTSTATUS TCPListen(PCONNECTION_ENDPOINT Connection, UINT Backlog)
+NTSTATUS TCPListen(PCONNECTION_ENDPOINT Connection, ULONG Backlog)
 {
     NTSTATUS Status = STATUS_SUCCESS;
     ip_addr_t AddressToBind;
